+++ test/test-diagnostics-channel-has-subscribers.spec.js
@@ -19,19 +19,7 @@ test('test-diagnostics-channel-has-subscribers', t => {
   const dc = channel('test-diagnostics-channel-has-subscribers');
   t.ok(!hasSubscribers('test-diagnostics-channel-has-subscribers'));
 
-  const handler = () => {};
-  dc.subscribe(handler);
+  dc.subscribe(() => {});
   t.ok(hasSubscribers('test-diagnostics-channel-has-subscribers'));
-
-  dc.unsubscribe(handler);
-  t.ok(!hasSubscribers('test-diagnostics-channel-has-subscribers'));
-
-  const store = {};
-  dc.bindStore(store);
-  t.ok(hasSubscribers('test-diagnostics-channel-has-subscribers'));
-
-  dc.unbindStore(store);
-  t.ok(!hasSubscribers('test-diagnostics-channel-has-subscribers'));
-
   t.end();
 });
